+++ database/migrations/2026_08_03_191340_create_attributes_table.php
@@ -17,11 +17,12 @@ return new class extends Migration
             // Flags
             $table->boolean('is_active')->default(false);
             $table->boolean('show_in_facets')->default(false);
-            $table->integer('sort_order')->default(1);
+            $table->unsignedBigInteger('sort_order')->default(1);
             // Descriptions
             $table->jsonb('name')->nullable()->default('{}');
             $table->timestamps();
 
+            // Indexes
             $table->index(['store_id', 'is_active', 'sort_order']);
         });
     }
