+++ database/migrations/2026_08_03_191949_create_attribute_values.php
@@ -19,7 +19,7 @@ return new class extends Migration
             // Flags
             $table->boolean('is_active')->default(false);
             $table->boolean('show_in_facets')->default(false);
-            $table->integer('sort_order')->default(1);
+            $table->unsignedBigInteger('sort_order')->default(1);
             // Descriptions
             $table->jsonb('name')->nullable()->default('{}');
             $table->jsonb('description')->nullable()->default('{}');
@@ -27,6 +27,7 @@ return new class extends Migration
             $table->tinyText('robots')->default('noindex, nofollow');
             $table->timestamps();
 
+            // Indexes
             $table->index(['attribute_id', 'store_id', 'is_active', 'sort_order']);
         });
     }
