+++ database/migrations/2026_08_06_003205_create_options_table.php
@@ -17,12 +17,13 @@ return new class extends Migration
             // Flags
             $table->boolean('is_active')->default(false);
             $table->boolean('show_in_facets')->default(false);
-            $table->integer('sort_order')->default(1);
+            $table->unsignedBigInteger('sort_order')->default(1);
             // Descriptions
             $table->jsonb('name')->nullable()->default('{}');
             $table->tinyText('type')->default('radio');
             $table->timestamps();
 
+            // Indexes
             $table->index(['store_id', 'is_active', 'sort_order']);
         });
     }
