+++ database/migrations/2026_08_06_003231_create_option_values_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->boolean('is_active')->default(false);
             $table->boolean('show_in_facets')->default(false);
             $table->boolean('is_default')->default(false);
-            $table->integer('sort_order')->default(1);
+            $table->unsignedBigInteger('sort_order')->default(1);
             // Descriptions
             $table->jsonb('name')->nullable()->default('{}');
             $table->jsonb('description')->nullable()->default('{}');
@@ -28,6 +28,7 @@ return new class extends Migration
             $table->tinyText('robots')->default('noindex, nofollow');
             $table->timestamps();
 
+            // Indexes
             $table->index(['option_id', 'store_id', 'is_active', 'sort_order']);
         });
     }
